renamed files according to structure
